+++ src/collective/solr/events.py
@@ -3,7 +3,11 @@ from Products.CMFCore.utils import getToolByName
 from Products.CMFDefault.permissions import ModifyPortalContent
 
 from Products.CMFCore.CMFCatalogAware import CMFCatalogAware
-from Products.Archetypes.CatalogMultiplex import CatalogMultiplex
+try:  # pragma: no cover
+    from Products.Archetypes.CatalogMultiplex import CatalogMultiplex
+except ImportError:  # pragma: no cover
+    class CatalogMultiplex(object):
+        pass
 
 
 def reorderedEvent(event):
@@ -11,6 +15,5 @@ def reorderedEvent(event):
     mtool = getToolByName(parent, 'portal_membership')
     if mtool.checkPermission(ModifyPortalContent, parent):
         for child in parent.objectValues():
-            if isinstance(child, CatalogMultiplex) or \
-                    isinstance(child, CMFCatalogAware):
+            if isinstance(child, (CatalogMultiplex, CMFCatalogAware)):
                 child.reindexObject(['getObjPositionInParent'])