+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 6.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Avoid hard/undeclared dependency on Archetypes
+  [tomgross]
 
 
 6.0a1.post1 (2017-05-22)